Implement timsort
